backend/mwalimu_boney_backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.4
cbor2==5.7.1
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
Django==6.0
django-cors-headers==4.5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dlib==20.0.0
docx==0.2.4
exceptiongroup==1.3.1
face-recognition==1.3.0
face_recognition_models==0.3.0
GDAL==3.8.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.1
lxml==6.0.2
msgpack==1.1.2
numpy==2.3.5
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python3-openid==3.2.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
zope.interface==8.1.1
